Fix: detect IPython and require magic for imports
On IPython, the kernel adds $PWD to sys.path before proceeding
to import lots of modules. In order to remove the restriction
that user submodule notebook names not shadow existing libraries
we can remove $PWD from the sys.path under IPython.
Once the kernel has started, experienced users *might* expect $PWD
to be on sys.path. Literary only restores this for non-__package__
notebooks.  This would also work out-of-the-box for top-level
notebooks, but not inside e.g. examples/ folders.